src/mlops/requirements.txt
pandas
numpy
scikit-learn
minio
mlflow
prometheus-client
hydra-core
slack_sdk
